% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WACSreadAgroclim.R
\name{WACSreadAgroclim}
\alias{WACSreadAgroclim}
\title{Function that reads a file of format Agroclim}
\usage{
WACSreadAgroclim(filename)
}
\arguments{
\item{filename}{path to the filemane to read}
}
\value{
A list of 2 objects:
        \itemize{
          \item{data}{The table data}
          \item{mapping}{The names of special var: year, month, day rain,
                         tmin and tmax}
        }
}
\description{
Function that reads a file of format Agroclim
}
\examples{
\dontrun{
  # using an agroclim format file
  WACSdata(WACSreadAgroclim("myfile.csv"))
}
}
