% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimateSeries.R
\docType{data}
\name{ClimateSeries}
\alias{ClimateSeries}
\title{Synthetic climate series of a french town between 1995 and 2012.}
\description{
year : Year
}
\details{
month : Month

day : Day

ETPP : Evaporation and sweating (mm)

RG : Radiance (Joules/cm²)

rain : Height of precipitation (mm)

tmin : Minimum temperature (Celsius)

tmax : Maximum temperature (Celsius)

V : Wind speed (meters per second)
}

